// frontend/src/api/clients/system/response.ts
type System = {
  data: {
    availableSeats: number[];
    bookedSeats: number[];
    reservedSeats: number[];
    bankVerifyStatus: boolean;
    bankAccount: string;
    bankAccountName: string;
    lastDBBackup: string;
    createdAt: string;
    updatedAt: string;
    sessionApiKey: string;
  };
};

export type { System };
export default null;